+++ manifests/init_scripts/lib/install_collective.sh
@@ -6,6 +6,11 @@
 #
 . /home/vagrant/manifests/init_scripts/lib_functions.sh
 
+serviceAction zookeeper stop
+serviceAction mesos-master stop
+serviceAction mesos-slave stop
+serviceAction marathon stop
+
 # Install Mesos
 bash manifests/init_scripts/lib/install_mesos.sh
 
@@ -15,7 +20,7 @@ bash manifests/init_scripts/lib/install_marathon.sh
 # # Install Mesos-docker-executor
 bash manifests/init_scripts/lib/install_mesos_docker.sh
 
-serviceRestart zookeeper
-serviceRestart mesos-master
-serviceRestart mesos-slave
-serviceRestart marathon
+serviceAction zookeeper restart
+serviceAction mesos-master restart
+serviceAction mesos-slave restart
+serviceAction marathon restart