+++ manifests/init_scripts/lib/install_mesos_docker.sh
@@ -16,6 +16,16 @@ if [ /var/lib/mesos/executors ]; then
   sudo mkdir -p /var/lib/mesos/executors
 fi
 
-echo "== Installing mesos-docker =="
-sudo cp -f /home/vagrant/manifests/repos/mesos-docker/bin/mesos-docker /var/lib/mesos/executors/docker
+function link() {
+  local src=/home/vagrant/manifests/repos/mesos-docker/bin/mesos-docker
+  local dest=/var/lib/mesos/executors/docker
 
+  if [ ! -e $dest ]; then
+    echo "== Installing mesos-docker... =="
+    sudo cp -f $src $dest
+  else
+    echo "== Mesos-docker already installed, skipping... =="
+  fi
+}
+
+link